refactor: apply code reivew
